imprv: Moved url related configs & utils to urls
