+++ lib/features/splash/mns_process/presentation/mns_query/mns_query_presenter.dart
@@ -1,5 +1,4 @@
 import 'package:datadashwallet/common/common.dart';
-import 'package:datadashwallet/common/config.dart';
 import 'package:datadashwallet/features/dapps/dapps.dart';
 import 'package:datadashwallet/core/core.dart';
 import 'package:flutter/material.dart';
@@ -111,8 +110,8 @@ class SplashMNSQueryPresenter extends CompletePresenter<SplashMNSQueryState> {
 
   Future<void> claim(String name) async {
     final launchUrl = state.network!.chainId == Config.mxcMainnetChainId
-        ? Config.mainnetMns(name)
-        : Config.testnetMns(name);
+        ? Urls.mainnetMns(name)
+        : Urls.testnetMns(name);
     await navigator
         ?.push(route.featureDialog(OpenAppPage(url: launchUrl)))
         .then((_) {
